⚡️ Car 객체 움직임에 전략패턴 도입
