+++ java-racingcar-master/README.md
@@ -54,7 +54,7 @@ int number = scanner.nextInt();
 ## 프로그래밍 요구사항
 - 모든 로직에 단위 테스트를 구현한다. 단 UI(System.out.System.in) 로직은 제외
   - 핵심 로직을 구현하는 코드와 UI 를 담당하는 로직을 구분한다.
-  - UI 로직을 InputView, ResultView 와 같은 클래스를 추가해 분리한다.
+  - UI 로직을 ui.InputView, ui.ResultView 와 같은 클래스를 추가해 분리한다.
 - 자바 코드 컨벤션을 지키면서 프로그래밍 한다.
   - 이 과정의 Code Style 은 Intellij idea Code Style Java 를 따른다.
   - Intellij idea Code Style Java 를 따르려면 code formatting 단축키 (Windows : Ctrl + Alt + L, Mac : (Option) + (Command) + L) 를 사용한다.
