--- java-racingcar-master/src/main/java/Car.java
@@ -1,16 +1,0 @@
-
-public class Car {
-
-    private int distance;
-
-    public void move(final boolean runnable) {
-        if(runnable){
-            distance = distance + 1;
-
-        }
-    }
-
-    public int getDistance() {
-        return this.distance;
-    }
-}
