+++ java-racingcar-master/src/main/java/Race.java
@@ -1,12 +1,10 @@
+import domain.Car;
+import domain.CarMoveAbleStrategyImpl;
+import domain.GameSet;
 import java.util.List;
-import java.util.Random;
+import ui.ResultView;
 
 public class Race {
-
-    private static final Random random = new Random();
-
-    private static final int NUMBER_OF_ADVANCES = 4;
-    private static final int RANDOM_BOUND = 10;
 
     public static void run(final GameSet gameSet) {
         for (int i = 0; i < gameSet.getCount(); i++) {
@@ -17,11 +15,7 @@ public class Race {
 
     public static void runGameRound(final List<Car> carList) {
         for (Car car : carList) {
-            car.move(isRun());
+            car.move(new CarMoveAbleStrategyImpl());
         }
-    }
-
-    private static boolean isRun() {
-        return random.nextInt(RANDOM_BOUND) > NUMBER_OF_ADVANCES;
     }
 }
