+++ java-racingcar-master/src/main/java/domain/Car.java
@@ -1,0 +1,14 @@
+package domain;
+
+public class Car {
+
+    private int distance;
+
+    public void move(final MoveAbleStrategy moveAbleStrategy) {
+        distance = distance + moveAbleStrategy.getMoveDistance();
+    }
+
+    public int getDistance() {
+        return this.distance;
+    }
+}
