+++ java-racingcar-master/src/main/java/domain/MoveAbleStrategy.java
@@ -1,0 +1,7 @@
+package domain;
+
+@FunctionalInterface
+public interface MoveAbleStrategy {
+
+    int getMoveDistance();
+}
