+++ java-racingcar-master/src/main/java/ui/InputView.java
@@ -1,3 +1,5 @@
+package ui;
+
 public class InputView {
 
     public void printCarsAsking() {
