+++ java-racingcar-master/src/main/java/ui/ResultView.java
@@ -1,3 +1,6 @@
+package ui;
+
+import domain.Car;
 import java.util.List;
 
 public class ResultView {
