--- java-racingcar-master/src/test/java/RaceTest.java
@@ -1,19 +1,0 @@
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-class RaceTest {
-
-    //todo 아래 메소드의 테스트를 어떻게 하쥬,,,????🔥
-    // private static boolean isRun() {
-    //        return random.nextInt(RANDOM_BOUND) > NUMBER_OF_ADVANCES;
-    //    }
-
-    @Test
-    @DisplayName("랜덤 숫자가 4이상일 때만 run 이 가능")
-    void isRunPossible() throws Exception {
-//        MockedStatic<Random> mockRandom = mockStatic(Random.class);
-//        Random random = new Random(9);
-//
-//        assertThat(Race.isRun()).isTrue();
-    }
-}
